Add printing of ast
Adds two functions, Ast.print_ast and Ast._print_ast - first one prints
ast in a beautified way, for example - (a . nil) will be printed as (a),
while second just prints everything and is meant as a debugging utility.

The way tokens are printed is also changed to match lisp syntax more.