Fix out-of-bounds error, add hash as a token
This mostly fixes indentation and error that would happen if indentifier
was the only thing in the input string. It's solved by simple bound
checking on every loop iteration inside identifier branch.

It also adds hash sign as a separate token for atoms such as #t (true)
or #f (false).

Adds nil to the Ast.