Add first, unfinished version of the evaluator
Adds evaluator and scope system, along with implementation of djb2 hash.
It also changes repl a little. Currently evaluator can bind values in a
scope and evaluate them.

From this point, all code is considered to be in the public domain, as
described in the UNLICENSE file.

Adds README with a small description of what the project is.